Extracted OpenIdConnectUserMustExist Handler into a separate class.
